+++ fabric/src/main/java/band/kessokuteatime/splasher/fabric/SplasherFabric.java
@@ -1,0 +1,11 @@
+package band.kessokuteatime.splasher.fabric;
+
+import band.kessokuteatime.splasher.Splasher;
+import net.fabricmc.api.ClientModInitializer;
+
+public final class SplasherFabric implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+		Splasher.config();
+	}
+}